requirements.txt
flask
flask_cors
fastapi
uvicorn
flask_httpaut